# server/src/routes/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API for demo project role based JWT authorization with Express and Vue ",
    "description": "Simple API for demo project by Mikhail Shpakov",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "Mikhail Shpakov",
      "email": "mikhail.n.shpakov@gmail.com",
      "url": "https://shpakov.dev"
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "login",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User authorization",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "description": "Username for user account",
                  "type": "string"
                },
                "password": {
                  "description": "Password for user account",
                  "type": "string"
                },
                "fingerprint": {
                  "description": "Browser fingerprint",
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success auth",
            "examples": {
              "application/json": {
                "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJlYTU2MGVhZC1mMWJkLTRiOWMtYTNhYS02YWEzNzlhYmQ2MjUiLCJwZXJtaXNzaW9ucyI6WyJhZG1pbiJdLCJpYXQiOjE1ODU3Nzg5OTgsImV4cCI6MTU4NTc5MzM5OH0.c07m2Z7LkwSfJLYY_HjE4bimWwyhZqXZVo7s9-a_70o",
                "refreshToken": "68ac980f-ca12-42ae-9373-6cb94e801968"
              }
            }
          },
          "401": {
            "description": "User with such credentials not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/auth/refresh-tokens": {
      "post": {
        "operationId": "refresh-tokens",
        "summary": "refresh-tokens",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User refresh tokens",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "description": "Refresh token that was stored on the frontend (uuid)",
                  "type": "string"
                },
                "fingerprint": {
                  "description": "Browser fingerprint",
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Tokens was success updated",
            "examples": {
              "application/json": {
                "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJlYTU2MGVhZC1mMWJkLTRiOWMtYTNhYS02YWEzNzlhYmQ2MjUiLCJwZXJtaXNzaW9ucyI6WyJhZG1pbiJdLCJpYXQiOjE1ODU3Nzg5OTgsImV4cCI6MTU4NTc5MzM5OH0.c07m2Z7LkwSfJLYY_HjE4bimWwyhZqXZVo7s9-a_70o",
                "refreshToken": "68ac980f-ca12-42ae-9373-6cb94e801968"
              }
            }
          },
          "401": {
            "description": "User with such credentials not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}